requirements.txt
Flask
flask-cors
eth-account
google-auth
python-dotenv
